+++ teste2SH/src/br/com/SistemaHospitalar/classes/EnfermeiraChefe.java
@@ -4,10 +4,10 @@ import br.com.SistemaHospitalar.enums.Turno;
 
 public class EnfermeiraChefe extends Enfermeira {
 
-    private Integer id_enfermeira_chefe;
+    private Long id_enfermeira_chefe;
 
-    public EnfermeiraChefe(Integer id, Ala ala, String cre, Turno enfermeiraTurno, String nome_enfermeira,
-            Integer id_enfermeira_chefe) {
+    public EnfermeiraChefe(Long id, Ala ala, String cre, Turno enfermeiraTurno, String nome_enfermeira,
+    		Long id_enfermeira_chefe) {
         super(id, ala, cre, enfermeiraTurno, nome_enfermeira);
         this.id_enfermeira_chefe = id_enfermeira_chefe;
     }
@@ -17,7 +17,7 @@ public class EnfermeiraChefe extends Enfermeira {
         return "EnfermeiraChefe [id_enfermeira_chefe=" + id_enfermeira_chefe + "]";
     }
 
-    public Integer getId_enfermeira_chefe() {
+    public Long getId_enfermeira_chefe() {
         return id_enfermeira_chefe;
     }
