+++ teste2SH/src/br/com/SistemaHospitalar/classes/Paciente.java
@@ -1,12 +1,12 @@
 package br.com.SistemaHospitalar.classes;
 
 public class Paciente {
-    private Integer id;
+    private Long id;
     private String CPF;
     private String nome;
     Medico medico;
 
-    public Paciente(Integer id, String cPF, String nome, Medico medico) {
+    public Paciente(Long id, String cPF, String nome, Medico medico) {
         this.id = id;
         CPF = cPF;
         this.nome = nome;
@@ -18,7 +18,7 @@ public class Paciente {
         return "Paciente id=" + id + ", CPF=" + CPF + ", nome=" + nome + ", medico=" + medico;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
